Added update note functionality, modified some handlers, moved from post to delete or get in some handlers and optimized some stuff

+++ src/db/models.rs
@@ -13,7 +13,7 @@ pub struct User {
 pub struct Notes {
     pub note_id: String,
     pub title: String,
-    pub priority: u8,
+    pub priority: u32,
     pub text: String,
     pub user: String,
     pub date: mongodb::bson::DateTime,
